% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_metadata.R
\name{core_metadata}
\alias{core_metadata}
\title{Core metadata}
\usage{
core_metadata(name = c())
}
\arguments{
\item{name}{Name of the core, given using quotes. Use preferably the published name of the core for traceability.}
}
\description{
Question the user on core International Geo Sample Number, sampling number, etc.
}
\examples{
user_infos()

}
\keyword{age-depth}
\keyword{modelling}
\keyword{visualisation}
