% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_serac.R
\name{help_serac}
\alias{help_serac}
\title{serac help}
\usage{
help_serac()
}
\description{
Give a help message
}
\examples{
serac_input_formatting('PB06')

}
\keyword{age-depth}
\keyword{modelling}
\keyword{visualisation}
