% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pckTest.R
\name{pkgTest}
\alias{pkgTest}
\title{Test package availability}
\usage{
pkgTest(x)
}
\description{
Question the user on core International Geo Sample Number, sampling number, etc.
}
\examples{
pkgTest('vegan')

}
