% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serac_example_ALO09P12_proxy.R
\docType{data}
\name{serac_example_ALO09P12_proxy}
\alias{serac_example_ALO09P12_proxy}
\title{Lake Allos shortlived proxy data}
\format{
rda file
}
\source{
\href{https://www.sciencedirect.com/science/article/pii/S0033589412000294}{Science Direct}
}
\usage{
serac_example_ALO09P12_proxy
}
\description{
Example of data used for serac
}
\examples{
# Radionuclides data
View(serac_example_ALO09P12)

# proxy data
View(serac_example_ALO09P12_proxy)

# Go to serac
# e.g.
serac("serac_example_ALO09P12", coring_yr = 2009)
}
\references{
Wilhelm, B., et al. (2012) Quat. Res. 78, 1–12.
(\href{https://www.cambridge.org/core/journals/quaternary-research/article/1400-years-of-extreme-precipitation-patterns-over-the-mediterranean-french-alps-and-possible-forcing-mechanisms/D6DBE2AD889B14347B540CD4CAB0C6E2}{Cambridge Core})
}
\keyword{datasets}
