% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serac_input_formatting.R
\name{serac_input_formatting}
\alias{serac_input_formatting}
\title{Prepare data input file}
\usage{
serac_input_formatting(name)
}
\arguments{
\item{name}{Name of the core, given using quotes. Use preferably the published name of the core for traceability.}
}
\description{
Function to prepare data input file. The input data file can easily be prepared in excel or other programs. The function below may help the user in checking the right input format. Function only for the general data file, 'varves' and 'proxy' input data have to be formated separately.
}
\examples{
serac_input_formatting('PB06')

}
\keyword{visualisation}
