% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_infos.R
\name{user_infos}
\alias{user_infos}
\title{Information on the user}
\usage{
user_infos()
}
\description{
Question the user on professional ID (name, ORCID, email, Institute), the first time serac is used.. These metadata are then included in the general metadata file output every time a model is created.
}
\examples{
user_infos()

}
\keyword{age-depth}
\keyword{data}
\keyword{management}
\keyword{modelling}
\keyword{reproducibility}
\keyword{visualisation}
