\name{core_metadata}
\alias{core_metadata}
\title{Core metadata}
\usage{
core_metadata(name="")
}
\description{
Question the user on professional ID (name, ORCID, email, Institute). These metadata are then included in the general metadata file output every time a model is created.
}
\examples{
core_metadata()
}
