% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mround.R
\name{mround}
\alias{mround}
\title{Round values}
\usage{
mround(x, base)
}
\arguments{
\item{x}{number to round up}

\item{base}{round up exponant}
}
\description{
Round value at the upper or lower value, defined by 'base'. Used for plotting purpose.
}
\examples{
mround(142,10)

}
\keyword{visualisation}
