\name{serac_input_formatting}
\alias{serac_input_formatting}
\title{Prepare input data file for serac use}
\usage{
serac_input_formatting(name)
}
\description{
Function to prepare data input file. The input data file can easily be prepared in excel or other programs. The function below may help the user in checking the right input format. Function only for the general data file, 'varves' and 'proxy' input data have to be formated separately.
}
\arguments{
\item{name}{Name of the core}
}
\examples{
serac_input_formatting(name = 'PB06')
}
