% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serac_map.R
\name{serac_map}
\alias{serac_map}
\title{Map the sites}
\usage{
serac_map(
  which_lakes = NULL,
  output_name = NULL,
  lon_lim = NULL,
  lat_lim = NULL,
  add_labels = TRUE
)
}
\arguments{
\item{which_lakes}{Name of the sites to be plotted. Must match folders names. Default: all the folder in the Cores folder.}

\item{output_name}{Extension to be added to the output name.}

\item{lon_lim}{Longitude limit for the plot (WGS84), expressed in a vector of length==2}

\item{lat_lim}{Latitude limit for the plot (WGS84), expressed in a vector of length==2}

\item{add_labels}{Logical argument. Whether the names of the sites (which_lakes) should be plotted.}
}
\description{
Function to map the different sites an user worked on.
}
\examples{
serac_input_formatting('PB06')

}
\keyword{age-depth}
\keyword{modelling}
\keyword{visualisation}
